+++ test/__init__.py
@@ -69,7 +69,16 @@ class TestMonitors(unittest.TestCase):
 
     def test_pods_monitor_output(self):
         gold = hashlib.md5(open("test/files/test_pods_are_stuck_pending.tf").read()).hexdigest()
-        new =hashlib.md5(open("test_pods_are_stuck_pending.tf").read()).hexdigest()
+        new = hashlib.md5(open("test_pods_are_stuck_pending.tf").read()).hexdigest()
+
+        logging.debug(gold)
+        logging.debug(new)
+
+        self.assertEqual(gold, new)
+
+    def test_increase_in_network_errors(self):
+        gold = hashlib.md5(open("test/files/test_increase_in_network_errors.tf").read()).hexdigest()
+        new = hashlib.md5(open("test_increase_in_network_errors.tf").read()).hexdigest()
 
         logging.debug(gold)
         logging.debug(new)
@@ -80,3 +89,4 @@ class TestMonitors(unittest.TestCase):
     def tearDown(self):
         os.remove("test_increase_in_network_bytes_received.tf")
         os.remove("test_pods_are_stuck_pending.tf")
+        os.remove("test_increase_in_network_errors.tf")