+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+## Updated:
+- added template definitions to the thresholds block of the monitors
+- added per monitor default definitions
+
 ## [1.8.1]
 ### Fixed
 - Fixed a typo in files/monitor.tf.jinja `evaluation_delay` variable name.